double precision variables

+++ source/neat.f95
@@ -1078,10 +1078,10 @@ contains
         DOUBLE PRECISION :: temp4
         LOGICAL, intent(in) :: norp
 
-        REAL :: fn_val
-        REAL     :: s = 0.449871, t = -0.386595, a = 0.19600, b = 0.25472,           &
+        double precision :: fn_val
+        double precision     :: s = 0.449871, t = -0.386595, a = 0.19600, b = 0.25472,           &
                     r1 = 0.27597, r2 = 0.27846, u, v, x, y, q
-        REAL :: half
+        double precision :: half
         double precision :: newmean, newsnr, snr
 
         half = 0.5
